refactor: Move components to components/common
