+++ src/__tests__/components/Header.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import { test, expect } from "vitest";
 
-import { Header } from "@/components/Header";
+import { Header } from "@/components/common/Header";
 
 test("match texts", () => {
   const props = {
