+++ src/app/page.tsx
@@ -8,10 +8,10 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Metadata } from "next";
 import Image from "next/image";
 
-import { Card } from "@/components/Card";
-import { Container } from "@/components/Container";
-import { Header } from "@/components/Header";
-import { TitledSecton } from "@/components/TitledSection";
+import { Card } from "@/components/common/Card";
+import { Container } from "@/components/common/Container";
+import { Header } from "@/components/common/Header";
+import { TitledSecton } from "@/components/common/TitledSection";
 import { MY_NAME, TITLE } from "@/constants/global";
 
 export const metadata: Metadata = {
